+++ OrderService/src/main/java/com/example/orderservice/global/config/RabbitMQConfig.java
@@ -1,11 +1,5 @@
 package com.example.orderservice.global.config;
 
-
-import org.springframework.amqp.core.Binding;
-import org.springframework.amqp.core.BindingBuilder;
-import org.springframework.amqp.core.DirectExchange;
-import org.springframework.amqp.core.Queue;
-import org.springframework.amqp.core.TopicExchange;
 import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
@@ -28,24 +22,6 @@ public class RabbitMQConfig {
 
     @Value("${spring.rabbitmq.password}")
     private String rabbitmqPassword;
-
-    @Value("${rabbitmq.exchange.name}")
-    private String exchangeName;
-
-    @Bean
-    public Queue queue() {
-        return new Queue("product.queue");
-    }
-
-    @Bean
-    public DirectExchange exchange() {
-        return new DirectExchange(exchangeName);
-    }
-
-    @Bean
-    public Binding binding(Queue queue, DirectExchange exchange) {
-        return BindingBuilder.bind(queue).to(exchange).with("product.key");
-    }
 
     @Bean
     public ConnectionFactory connectionFactory() {
