+++ productService/src/main/java/com/example/productservice/ProductServiceApplication.java
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
-//@EnableScheduling
+@EnableScheduling
 public class ProductServiceApplication {
 
     public static void main(String[] args) {
