--- productService/src/main/java/com/example/productservice/global/dto/MessageDto.java
@@ -1,14 +1,0 @@
-package com.example.productservice.global.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class MessageDto {
-
-    private String title;
-    private String content;
-}
