+++ productService/src/main/java/com/example/productservice/product/dto/ProductDto.java
@@ -3,8 +3,6 @@ package com.example.productservice.product.dto;
 import com.example.productservice.product.entity.Status;
 import lombok.Data;
 
-import java.time.LocalDateTime;
-
 @Data
 public class ProductDto {
     private String productUUID;
