+++ productService/src/main/java/com/example/productservice/product/service/ProductService.java
@@ -2,7 +2,6 @@ package com.example.productservice.product.service;
 
 import com.example.productservice.product.dto.ProductDto;
 import com.example.productservice.product.entity.Content;
-import com.example.productservice.product.vo.ResponseProduct;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -14,7 +13,7 @@ public interface ProductService {
     Page<ProductDto> getProductByName(String productName, int pageNum, int pageSize);
 
     // internal 의 서비스
-    ResponseProduct increaseCount(Content content);
+    int increaseCount(Content content);
 
     int decreaseCount(Content content);
 }
