+++ productService/src/main/java/com/example/productservice/product/service/ProductServiceImpl.java
@@ -6,20 +6,14 @@ import com.example.productservice.product.entity.Product;
 import com.example.productservice.product.entity.Status;
 import com.example.productservice.product.mapstruct.ProductMapStruct;
 import com.example.productservice.product.repository.ProductRepository;
-import com.example.productservice.product.vo.ResponseProduct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.redisson.api.RLock;
-import org.redisson.api.RedissonClient;
-import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.concurrent.TimeUnit;
 
 @Slf4j
 @Service
@@ -28,10 +22,6 @@ public class ProductServiceImpl implements ProductService{
 
     private final ProductMapStruct productMapStruct;
     private final ProductRepository productRepository;
-    private final RedissonClient redissonClient;
-    private final RedisService redisService;
-
-    private int numbering;
 
     @Override
     public int createProduct(ProductDto productDto) {
@@ -70,34 +60,19 @@ public class ProductServiceImpl implements ProductService{
 
     // internal 의 서비스
     @Override
-    public ResponseProduct increaseCount(Content content) {
-        RLock lock = redissonClient.getLock(content.getProductUUID());
+    @Transactional
+    public int increaseCount(Content content) {
+        Product product = productRepository.findByProductUUIDForUpdate(content.getProductUUID());
 
-        try{
-            boolean available = lock.tryLock(5, 2, TimeUnit.SECONDS);
-            if (!available) {
-                throw new RuntimeException("Lock 획득 실패!");
-            }
-            Product product = productRepository.findByProductUUID(content.getProductUUID());
-
-            if (product == null) {
-                throw new IllegalArgumentException("제품을 찾을 수 없습니다");
-            }
-
-            if (product.getStock() < content.getUnitCount()) {
-                throw new RuntimeException("재고가 부족하여 주문을 생성할 수 없습니다");
-            }
-
-            product.increaseStock(content.getUnitCount());
-
-            productRepository.save(product);
-
-            return productMapStruct.changeResponse(productMapStruct.changeDto(product));
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        } finally {
-            lock.unlock();
+        if (product == null) {
+            throw new RuntimeException("상품이 존재하지 않습니다");
         }
+
+        product.increaseStock(content.getUnitCount());
+
+        productRepository.save(product);
+
+        return 1;
     }
 
     @Override
