+++ productService/src/main/java/com/example/productservice/product/service/RedisService.java
@@ -3,8 +3,4 @@ package com.example.productservice.product.service;
 public interface RedisService {
 
     void setValue(String key, String value);
-
-    void setRemove(String key);
-
-    Long countingCard(String key);
 }
