+++ productService/src/main/java/com/example/productservice/product/service/RedisServiceImpl.java
@@ -2,7 +2,6 @@ package com.example.productservice.product.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -15,16 +14,4 @@ public class RedisServiceImpl implements RedisService {
     public void setValue(String key, String value) {
         redisTemplate.opsForSet().add(key, value);
     }
-
-    @Override
-    public void setRemove(String key) {
-        redisTemplate.opsForSet().remove(key);
-    }
-
-    @Override
-    public Long countingCard(String key) {
-        return redisTemplate.opsForSet().size(key);
-    }
-
-
 }
