+++ productService/src/main/java/com/example/productservice/product/vo/RequestProduct.java
@@ -2,9 +2,6 @@ package com.example.productservice.product.vo;
 
 import lombok.Data;
 
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
 @Data
 public class RequestProduct {
     private String name;
