+++ userService/src/main/java/com/example/userservice/member/service/MemberService.java
@@ -6,8 +6,6 @@ import com.example.userservice.member.vo.RequestPassword;
 import com.example.userservice.member.vo.ResponseMember;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
-import jakarta.validation.Valid;
-import org.springframework.web.bind.annotation.RequestBody;
 
 public interface MemberService {
     String createMember(MemberDto memberDto, HttpSession session);
