+++ userService/src/main/java/com/example/userservice/member/service/RedisService.java
@@ -3,11 +3,9 @@ package com.example.userservice.member.service;
 import java.time.Duration;
 
 public interface RedisService {
-    void setValues(String key, String value);
 
     void setValues(String key, String value, Duration duration);
 
     String getValue(String key);
 
-    void deleteValue(String key);
 }
