+++ userService/src/main/java/com/example/userservice/member/vo/RequestCheck.java
@@ -1,6 +1,5 @@
 package com.example.userservice.member.vo;
 
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
