+++ userService/src/main/java/com/example/userservice/member/vo/ResponseMember.java
@@ -3,7 +3,6 @@ package com.example.userservice.member.vo;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 
-import java.util.List;
 
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
